ENH: Add spin and automatic calibration and to vtkIGSIOPivotCalibrationAlgo
This commit adds functions for performing spin calibration to vtkIGSIOPivotCalibrationAlgo using the same methods as pivot calibration.
This commit also adds the  ability to use the pivot/spin calibration algorithms continuously.

When AutoCalibrationEnabled is true, input poses will be placed into buckets. Once the latest bucket is filled, calibration is performed on the bucket to ensure that it contains pivoting/spinning motion. If the calibration error within the bucket is too high it means that the stylus is not pivoting/spinning, and all buckets are discarded.
Once the desired number of points has been collected, then calibration is run every time a bucket is filled. If the error for the calibration is less than the threshold, the result will be saved, and AutoCalibrationCompleteEvent will be invoked.

Default behavior should not have changed.